Merge pull request #82 from duniyalr/master

+++ src/utils/calculations/Rotation.java
@@ -3,12 +3,11 @@ package utils.calculations;
 import java.lang.Math;
 
 import config.Config;
-import elementManager.ElementManager;
 import elementManager.coordinate.AaPosition;
 import elementManager.elements.*;
 
 public class Rotation{
-    private Integer rotationSpeed; //should be in degree
+    private double rotationSpeed; //should be in degree
     private AaPosition origin; //coordinates of center of the Big Ball
     private Integer r; // radius of the Main Circle
 
@@ -16,11 +15,11 @@ public class Rotation{
         this.rotationSpeed = 2;
     }
 
-    public Rotation(Integer rotationSpeed){
+    public Rotation(double rotationSpeed){
         this.rotationSpeed = rotationSpeed;
     }
 
-    public void setSpeed(Integer rotationSpeed){
+    public void setSpeed(double rotationSpeed){
         this.rotationSpeed = rotationSpeed;
     }
